+++ packages/atom-react-redux/src/actionExtensions.ts
@@ -2,10 +2,16 @@ import { createAction, ActionCreatorWithPreparedPayload } from "@reduxjs/toolkit
 
 export type ActionResolveFunc<TResult> = (value?: TResult) => void;
 export type ActionRejectFunc = (reason?: string) => void;
+export type ActionValidationErrors<TFormData extends Record<string, unknown>> = Partial<Record<Extract<keyof TFormData, string>, string[]>>;
+export type ActionValidationRejectFunc<TFormData extends Record<string, unknown>> = (errors: ActionValidationErrors<TFormData>) => void;
 
 export interface ActionMeta<TResult> {
     resolve?: ActionResolveFunc<TResult>;
     reject?: ActionRejectFunc;
+}
+
+export interface ActionMetaValidatable<TResult, TFormData extends Record<string, unknown>> extends ActionMeta<TResult> {
+    validationReject?: ActionValidationRejectFunc<TFormData>;
 }
 
 /**
@@ -23,9 +29,41 @@ export function createActionWithMeta<TPayload = void, TResult = void>(type: stri
 > {
     return createAction(
         type,
-        (payload: TPayload, resolve: ActionResolveFunc<TResult> = () => void 0, reject: ActionRejectFunc = () => void 0) => ({
+        (
+            payload: TPayload,
+            resolve: ActionResolveFunc<TResult> = () => void 0,
+            reject: ActionRejectFunc = () => void 0
+        ) => ({
             payload,
             meta: { resolve, reject },
+        })
+    );
+}
+
+/**
+ * Creates an action which accepts a payload, optional validation callback, and optional resolve/reject callbacks as meta.
+ */
+export function createActionWithMetaValidatable<
+    TPayload = void,
+    TResult = void,
+    TFormData extends Record<string, unknown> = Record<string, unknown>
+>(type: string): ActionCreatorWithPreparedPayload<
+    [TPayload, ActionValidationRejectFunc<TFormData>?, ActionResolveFunc<TResult>?, ActionRejectFunc?],
+    TPayload,
+    string,
+    never,
+    ActionMetaValidatable<TResult, TFormData>
+> {
+    return createAction(
+        type,
+        (
+            payload: TPayload,
+            validationReject: ActionValidationRejectFunc<TFormData> = () => void 0,
+            resolve: ActionResolveFunc<TResult> = () => void 0,
+            reject: ActionRejectFunc = () => void 0,
+        ) => ({
+            payload,
+            meta: { validationReject, resolve, reject },
         })
     );
 }