+++ packages/atom-react-redux/src/index.ts
@@ -7,11 +7,19 @@ export { setNotificationService, getNotificationService } from "./notifications"
 export type { NotificationService } from "./notifications";
 
 // Action extensions
-export { createActionWithMeta } from "./actionExtensions";
-export type { ActionMeta, ActionResolveFunc, ActionRejectFunc } from "./actionExtensions";
+export { createActionWithMeta, createActionWithMetaValidatable } from "./actionExtensions";
+export type {
+    ActionMeta,
+    ActionResolveFunc,
+    ActionRejectFunc,
+    ActionMetaValidatable,
+    ActionValidationErrors,
+    ActionValidationRejectFunc,
+} from "./actionExtensions";
 
 // Saga utilities
-export { withLoading, withCallback } from "./utils";
+export { withLoading, withCallback, withValidatableCallback } from "./utils";
+export type { ValidatableCallbackOptions } from "./utils";
 
 // callApi saga builder
 export { callApi } from "./callApi";